scripts/train.py
#!/usr/bin/env python3
"""
Training Script for SAR Autoencoder

Usage:
    python scripts/train.py --config configs/default.yaml
    python scripts/train.py --data patches.npy --epochs 50 --latent_channels 64
"""

import argparse
import yaml
from pathlib import Path
import sys

# Add src to path
sys.path.insert(0, str(Path(__file__).parent.parent))

from src.models import SARAutoencoder
from src.data import SARDataModule
from src.losses import CombinedLoss
from src.training import Trainer


def parse_args():
    parser = argparse.ArgumentParser(description='Train SAR Autoencoder')
    
    # Data
    parser.add_argument('--data', type=str, default='data/patches/patches.npy',
                        help='Path to patches file')
    parser.add_argument('--val_fraction', type=float, default=0.1,
                        help='Validation fraction')
    
    # Model
    parser.add_argument('--latent_channels', type=int, default=64,
                        help='Latent channels (controls compression)')
    parser.add_argument('--base_channels', type=int, default=64,
                        help='Base channel count')
    
    # Training
    parser.add_argument('--epochs', type=int, default=50,
                        help='Training epochs')
    parser.add_argument('--batch_size', type=int, default=16,
                        help='Batch size')
    parser.add_argument('--lr', type=float, default=1e-4,
                        help='Learning rate')
    parser.add_argument('--mse_weight', type=float, default=1.0,
                        help='MSE loss weight')
    parser.add_argument('--ssim_weight', type=float, default=0.1,
                        help='SSIM loss weight')
    
    # Config file (overrides other args)
    parser.add_argument('--config', type=str, default=None,
                        help='Path to config YAML file')
    
    # Output
    parser.add_argument('--checkpoint_dir', type=str, default='checkpoints',
                        help='Checkpoint directory')
    parser.add_argument('--log_dir', type=str, default='runs',
                        help='TensorBoard log directory')
    
    return parser.parse_args()


def main():
    args = parse_args()
    
    # Load config if provided
    if args.config:
        with open(args.config) as f:
            config = yaml.safe_load(f)
    else:
        config = vars(args)
    
    print("=" * 60)
    print("SAR Autoencoder Training")
    print("=" * 60)
    print(f"Configuration: {config}")
    
    # TODO: Implement training script
    #
    # 1. Create data module
    # data = SARDataModule(
    #     config['data'],
    #     val_fraction=config['val_fraction'],
    #     batch_size=config['batch_size'],
    # )
    #
    # 2. Create model
    # model = SARAutoencoder(
    #     latent_channels=config['latent_channels'],
    #     base_channels=config['base_channels'],
    # )
    #
    # 3. Create loss function
    # loss_fn = CombinedLoss(
    #     mse_weight=config['mse_weight'],
    #     ssim_weight=config['ssim_weight'],
    # )
    #
    # 4. Create trainer
    # trainer = Trainer(
    #     model=model,
    #     train_loader=data.train_dataloader(),
    #     val_loader=data.val_dataloader(),
    #     loss_fn=loss_fn,
    #     config=config,
    # )
    #
    # 5. Train
    # trainer.train(epochs=config['epochs'])
    
    print("\n[TODO: Implement training - see src/training/trainer.py]")


if __name__ == '__main__':
    main()
