requirements.txt
# Core
numpy
scipy
matplotlib
torch>=2.0.0
torchvision>=0.15.0

# Deep Learning
torch
torchvision

# Image Processing
Pillow
scikit-image

# SAR/Geospatial
rasterio>=1.3.0

# Compression
PyWavelets

# Jupyter
jupyter
ipykernel

# Experiment Tracking
tensorboard>=2.12.0
tqdm>=4.65.0
pyyaml>=6.0

# Utilities
tqdm
pyyaml
jupyter>=1.0.0

# Metrics
pytorch-msssim>=1.0.0